+++ assets/app-config.js
@@ -1,2 +1,2 @@
-var ASSET_VER = '7464';
+var ASSET_VER = '7465';
 window.ASSET_VER = ASSET_VER;